refactor and add e2e test
