+++ e2e/tests/mobile/guest-shopper.spec.js
@@ -7,6 +7,7 @@
 
 const { test, expect } = require("@playwright/test");
 const config = require("../../config");
+const { addProductToCartMobile } = require("../../scripts/pageHelpers");
 const {
   generateUserCredentials,
   getCreditCardExpiry,
@@ -15,74 +16,7 @@ const {
 const GUEST_USER_CREDENTIALS = generateUserCredentials();
 
 test("Guest shopper can checkout items as guest", async ({ page }) => {
-  // Home page
-  await page.goto(config.RETAIL_APP_HOME);
-
-  await page.getByLabel("Menu", { exact: true }).click();
-
-  // SSR nav loads top level categories as direct links so we wait till all sub-categories load in the accordion
-  const categoryAccordion = page.locator(
-    "#category-nav .chakra-accordion__button svg+:text('Womens')"
-  );
-  await categoryAccordion.waitFor();
-
-  await page.getByRole("button", { name: "Womens" }).click();
-
-  const clothingNav = page.getByRole("button", { name: "Clothing" });
-
-  await clothingNav.waitFor();
-
-  await clothingNav.click();
-
-  const topsLink = page.getByLabel('Womens').getByRole("link", { name: "Tops" })
-  await topsLink.click();
-  // Wait for the nav menu to close first
-  await topsLink.waitFor({state: 'hidden'})
-
-  await expect(page.getByRole("heading", { name: "Tops" })).toBeVisible();
-
-  // PLP
-  const productTile = page.getByRole("link", {
-    name: /Cotton Turtleneck Sweater/i,
-  });
-  await productTile.scrollIntoViewIfNeeded()
-  // selecting swatch
-  const productTileImg = productTile.locator("img");
-  await productTileImg.waitFor({state: 'visible'})
-  const initialSrc = await productTileImg.getAttribute("src");
-  await expect(productTile.getByText(/From \$39\.99/i)).toBeVisible();
-
-  await productTile.getByLabel(/Black/, { exact: true }).click();
-  // Make sure the image src has changed
-  await expect(async () => {
-    const newSrc = await productTileImg.getAttribute("src")
-    expect(newSrc).not.toBe(initialSrc)
-  }).toPass()
-  await expect(productTile.getByText(/From \$39\.99/i)).toBeVisible();
-  await productTile.click();
-
-  // PDP
-  await expect(
-    page.getByRole("heading", { name: /Cotton Turtleneck Sweater/i })
-  ).toBeVisible();
-  await page.getByRole("radio", { name: "L", exact: true }).click();
-
-  await page.locator("button[data-testid='quantity-increment']").click();
-
-  // Selected Size and Color texts are broken into multiple elements on the page.
-  // So we need to look at the page URL to verify selected variants
-  const updatedPageURL = await page.url();
-  const params = updatedPageURL.split("?")[1];
-  expect(params).toMatch(/size=9LG/i);
-  expect(params).toMatch(/color=JJ169XX/i);
-
-  await page.getByRole("button", { name: /Add to Cart/i }).click();
-
-  const addedToCartModal = page.getByText(/2 items added to cart/i);
-
-  await addedToCartModal.waitFor();
-
-  await page.getByLabel("Close").click();
+  await addProductToCartMobile({page})
 
   // Cart
   await page.getByLabel(/My cart/i).click();
@@ -180,4 +114,29 @@ test("Guest shopper can checkout items as guest", async ({ page }) => {
   await expect(
     page.getByRole("link", { name: /Cotton Turtleneck Sweater/i })
   ).toBeVisible();
+});
+
+test("Guest shopper can edit product item in cart", async ({ page }) => {
+  await addProductToCartMobile({page})
+
+  // Cart
+  await page.getByLabel(/My cart/i).click();
+
+  await expect(
+    page.getByRole("link", { name: /Cotton Turtleneck Sweater/i })
+  ).toBeVisible();
+
+  await expect(page.getByText(/Color: Black/i)).toBeVisible()
+  await expect(page.getByText(/Size: L/i)).toBeVisible()
+
+  await page.getByRole("button", { name: "Edit" }).click();
+  await expect(page.getByTestId('product-view')).toBeVisible()      
+  
+  // update variant in product edit modal
+  await page.getByRole("radio", { name: "S", exact: true }).click();
+  await page.getByRole("radio", { name: "Meadow Violet", exact: true }).click();
+  await page.getByRole("button", { name: /Update/i }).click()
+
+  await expect(page.getByText(/Color: Meadow Violet/i)).toBeVisible()
+  await expect(page.getByText(/Size: S/i)).toBeVisible()
 });
